// src/app/interfaces/IContasReceber.ts
export interface IContasReceber {
    codProduto: string;
    desEstabel: string;
    desGrupo: string;
    desSessao: string;
    numFaixaFinal: number;
    numFaixaInicial: number;
    numLitros: number;
    numServicos: number;
    numValor5dias: number;
    numValor10dias: number;
    numValor20dias: number;
    numValor30dias: number;
    numValorHoje: number;
    numValorMais30: number;
    numVencido5a9: number;
    numVencido10a19: number;
    numVencido20a29: number;
    numVencido30a59: number;
    numVencidoAte5: number;
    numVencidomais60: number;
    numfrete: number;
}